requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-dotenv==0.19.2
openai==0.11.0
opencv-python-headless==4.5.5.64